+++ interfaces/HO-Portal/src/app/components/registration-activity-overview/registration-activity-overview.component.ts
@@ -123,8 +123,6 @@ export class RegistrationActivityOverviewComponent implements OnInit {
       );
       this.pastTransactions = await this.programsService.getTransactions(
         this.program.id,
-        null,
-        null,
         this.person?.referenceId,
       );
       this.pastPayments = await this.programsService.getPastPayments(