+++ interfaces/HO-Portal/src/app/services/programs-service-api.service.ts
@@ -22,7 +22,7 @@ import {
 } from '../models/program.model';
 import { RegistrationChangeLog } from '../models/registration-change-log.model';
 import { RegistrationStatusChange } from '../models/registration-status-change.model';
-import { Transaction } from '../models/transaction.model';
+import { PaymentSummary, Transaction } from '../models/transaction.model';
 import { Role, TableData, User, UserSearchResult } from '../models/user.model';
 import { ImportResult } from '../program/bulk-import/bulk-import.component';
 import { arrayToXlsx } from '../shared/array-to-xlsx';
@@ -196,26 +196,26 @@ export class ProgramsServiceApiService {
 
   getTransactions(
     programId: number | string,
-    minPayment?: number | string,
-    payment?: number | string,
-    referenceId?: string,
+    referenceId: string,
   ): Promise<Transaction[]> {
     let params = new HttpParams();
-    if (minPayment) {
-      params = params.append('minPayment', minPayment);
-    }
-    if (payment) {
-      params = params.append('payment', payment);
-    }
-    if (referenceId) {
-      params = params.append('referenceId', referenceId);
-    }
+    params = params.append('referenceId', referenceId);
     return this.apiService.get(
       environment.url_121_service_api,
-      `/programs/${programId}/payments/transactions`,
+      `/programs/${programId}/transactions`,
       false,
       false,
       params,
+    );
+  }
+
+  getPaymentSummary(
+    programId: number | string,
+    payment: number | string,
+  ): Promise<PaymentSummary> {
+    return this.apiService.get(
+      environment.url_121_service_api,
+      `/programs/${programId}/payments/${payment}`,
     );
   }
 
@@ -291,7 +291,7 @@ export class ProgramsServiceApiService {
     }
     return this.apiService.get(
       environment.url_121_service_api,
-      `/programs/${programId}/payments/transactions/one`,
+      `/programs/${programId}/transactions/one`,
       false,
       false,
       params,