+++ services/121-service/src/main.ts
@@ -24,6 +24,7 @@ import {
   SWAGGER_CUSTOM_JS,
 } from './config';
 import appInsights = require('applicationinsights');
+import { AzureLogService } from './shared/services/azure-log.service';
 
 /**
  * A visualization of module dependencies is generated using `nestjs-spelunker`
@@ -69,16 +70,8 @@ function generateModuleDependencyGraph(app: INestApplication): void {
 async function bootstrap(): Promise<void> {
   const app = await NestFactory.create(ApplicationModule);
 
-  let corsAllowList: string[] | RegExp[];
-
-  if (!!process.env.CORS_ALLOW_LIST) {
-    corsAllowList = process.env.CORS_ALLOW_LIST.split(',').map(
-      (origin) => new RegExp(origin),
-    );
-  }
-
   app.enableCors({
-    origin: DEBUG ? true : corsAllowList || false,
+    origin: DEBUG,
     methods: 'GET,HEAD,PUT,PATCH,POST,DELETE,OPTIONS',
     credentials: true,
   });
@@ -170,6 +163,25 @@ async function bootstrap(): Promise<void> {
 
   const server = await app.listen(PORT);
   server.setTimeout(10 * 60 * 1000);
+
+  // Set up generic error handling:
+  process.on('unhandledRejection', (reason: string, promise: Promise<any>) => {
+    console.warn('Unhandled Rejection:', reason, promise);
+    throw reason;
+  });
+
+  process.on('uncaughtException', (error: Error) => {
+    console.warn('Uncaught Exception:', error);
+
+    const logService = new AzureLogService();
+    if (logService) {
+      logService.logError(error, true);
+      logService.logError(new Error('Uncaught Exception: restarting'), true);
+    }
+
+    // Trigger a reboot, as the app is in an unknown state.
+    process.exit(1);
+  });
 }
 void bootstrap();