+++ services/121-service/src/notifications/whatsapp/whatsapp.service.ts
@@ -202,7 +202,8 @@ export class WhatsappService {
       }
       await this.twilioMessageRepository.save(twilioMessage);
     }
-    await this.lastMessageService.updateLastMessageStatus(message.sid);
+    // This is commented out because we think this is causing performance issues
+    // await this.lastMessageService.updateLastMessageStatus(message.sid);
   }
 
   public async findOne(sid: string): Promise<TwilioMessageEntity> {