+++ services/121-service/src/payments/dto/get-payment-aggregration.dto.ts
@@ -1,0 +1,9 @@
+import { IsNumberString } from 'class-validator';
+
+export class GetPaymentAggregationDto {
+  @IsNumberString()
+  programId: number;
+
+  @IsNumberString()
+  payment: number;
+}