+++ services/121-service/src/scripts/seed-multiple-nlrc.ts
@@ -17,10 +17,9 @@ export class SeedMultipleNLRC implements InterfaceScript {
 
   private readonly seedHelper = new SeedHelper(this.dataSource);
 
-  public async run(): Promise<void> {
+  public async run(isApiTests?: boolean): Promise<void> {
     const seedInit = await new SeedInit(this.dataSource);
-    await seedInit.run();
-
+    await seedInit.run(isApiTests);
     // ************************
     // ***** Program LVV *****
     // ************************