+++ CHANGELOG.md
@@ -5,21 +5,96 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/global-121/121-platform/compare/v1.104.1...master)
+## [Unreleased](https://github.com/global-121/121-platform/compare/v1.105.9...master)
+
+## Removed
+
+- CORS configuration should be hanndled via the Azure Portal for each App Service instance.  
+  See: A step in the "How to set up a new instance"-guide in the wiki.
+
+---
+## [1.105.9](https://github.com/global-121/121-platform/compare/v1.105.8...v1.105.9) - 2023-11-15
+
+### Fixed
+
+- Filter for phonenumber of imported/invited PA
+
+
+## [1.105.8](https://github.com/global-121/121-platform/compare/v1.105.7...v1.105.8) - 2023-11-15
+
+### Added
+
+- Temporarily added twice-a-day cronjob to bulk update last message status
+
+## [1.105.7](https://github.com/global-121/121-platform/compare/v1.105.6...v1.105.7) - 2023-11-15
+
+### Fixed
+
+- Cannot include completed PA with 0 remaining payments (additional fix)
+
+## [1.105.6](https://github.com/global-121/121-platform/compare/v1.105.5...v1.105.6) - 2023-11-15
+
+### Fixed
+
+- Cannot include completed PA with 0 remaining payments
+
+## [1.105.5](https://github.com/global-121/121-platform/compare/v1.105.4...v1.105.5) - 2023-11-15
+
+### Fixed
+
+- paymentCount in PA popup
+- Setting of completed status & paymentCount
+
+## [1.105.4](https://github.com/global-121/121-platform/compare/v1.105.3...v1.105.4) - 2023-11-13
+
+### Fixed
+
+- Don't update last message status (performance issue) (additional places)
+
+## [1.105.3](https://github.com/global-121/121-platform/compare/v1.105.2...v1.105.3) - 2023-11-13
+
+### Fixed
+
+- Don't update last message status (performance issue)
+
+## [1.105.2](https://github.com/global-121/121-platform/compare/v1.105.1...v1.105.2) - 2023-11-07
+
+### Fixed
+
+- Change permissions for GET '/roles' endpoint
+
+## [1.105.1](https://github.com/global-121/121-platform/compare/v1.105.0...v1.105.1) - 2023-11-06
+
+### Added
+
+- Filter registrations in PA-table on 'Registration created date'
+
+## [1.105.0](https://github.com/global-121/121-platform/compare/v1.104.2...v1.105.0) - 2023-11-03
+
+### Fixed
+
+- Improved Commercial Bank of Ethiopia error handling
 
 ### Changed
 
-- Upgraded Node.js to v18 for PA-App + HO-Portal.  
+- Upgraded Node.js to v18 for PA-App + HO-Portal.
   Make sure to use your Node version manager to install+run the newest version.
-- Upgraded Node.js to v18 for 121-service.  
+- Upgraded Node.js to v18 for 121-service.
   Make sure to recreate the Docker image using `npm run start:services`.
+- Improved performance of 'last payment summary' in payment page
+- Refactored API-paths
+- Prohibit note-creation for referenceIds of programs you do not have permission for
 
 ### Removed
 
 - All audio-features are removed from the PA-App. To prevent a lot of "local changes" during development,
   make sure to remove any previously generated `*.webm`/`*.mp3` audio-files at: `/interfaces/PA-App/src/assets/<language-code>`
 
----
+## [1.104.2](https://github.com/global-121/121-platform/compare/v1.104.1...v1.104.2) - 2023-11-02
+
+### Fixed
+
+- Permission status change fix
 
 ## [1.104.1](https://github.com/global-121/121-platform/compare/v1.104.0...v1.104.1) - 2023-10-31